refactor: extract test fixtures
