+++ src/test/java/roomescape/fixture/TimeFixture.java
@@ -1,0 +1,19 @@
+package roomescape.fixture;
+
+import java.util.List;
+
+import roomescape.domain.reservation.ReservationTime;
+
+public class TimeFixture {
+
+    public static final List<ReservationTime> TIMES = List.of(
+            new ReservationTime(1L, "10:00"),
+            new ReservationTime(2L, "11:00"),
+            new ReservationTime(3L, "12:00"),
+            new ReservationTime(4L, "13:00")
+    );
+
+    public static ReservationTime timeFixture(int id) {
+        return TIMES.get(id - 1);
+    }
+}
